Add initial content
RoR app to be deployed on cPanel. The rails versions were
incompatibel so I had to just deploy/test it as a static site.

As a result dockerized a flow to launch it locally, scrape it,
then view/test the static site.

The host they use had to enable ssh access to allow this flow to
work.