BANG
Lots of iterations and changes. Should be everything needed to deploy
the site with Apcera.

Due to site using cPanel, deploying this Rails app was impossible. As
a hack I just scraped the static site and deployed it manually as such.

There were a few hacks I had to do to keep font-awesome and the
javascript working appropriately which I have locally.