Alternate branch: static site
Due to deploying to a cheap host; there were numerous problems when
trying to deploy an actual ruby app. The people I'm doing this for
have paid for more than a year so it wasn't reasonable to have them
dump their host at this time.

As an alternate deploy plan; I deployed the site locally then
scraped it so taht I could deploy it as a static site.

There were numerous changes that I had to make to fix the site
after that.

Some changes have occured to the static files that are no longer
reflected in the original RoR version.